+++ tsup.config.ts
@@ -3,6 +3,7 @@ import { defineConfig } from "tsup";
 export default defineConfig({
   entry: {
     index: "src/index.ts",
+    plugin: "src/plugin.ts",
     "core/index": "src/core/index.ts",
   },
   format: ["esm"],